refactor(vega): move safe SQL logging helper
