+++ adp/vega/vega-backend/server/logics/connector/local/table/safe_log.go
@@ -1,4 +1,4 @@
-package safelog
+package table
 
 import (
 	"crypto/sha256"
