+++ adp/vega/vega-backend/server/logics/connector/local/table/safe_log_test.go
@@ -1,4 +1,4 @@
-package safelog
+package table
 
 import (
 	"strings"
